### 2020-05-29 #2273 @danielskatz
👋 @adamgdobrakowski - Thanks for your submission to JOSS. As described in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss) announcing the reopening of JOSS, we're currently working in a "reduced service mode", limiting the number of papers assigned to any individual editor.

Since reopening JOSS last week, we've had > 60 papers submitted and as such, yours has been put in our backlog that we will be working through over the coming weeks and months.

Thanks in advance for your patience!

(In the meantime, feel free to make any needed fixes to the paper, such as adding the possibly missing DOI, etc. - add ```@whedon generate pdf``` as a new comment in this thread to build the pdf)
