+++ paper.bib
@@ -22,7 +22,8 @@
   author={Pennington, Jeffrey and Socher, Richard and Manning, Christopher},
   booktitle={Proceedings of the 2014 conference on empirical methods in natural language processing (EMNLP)},
   pages={1532--1543},
-  year={2014}
+  year={2014},
+  doi = {10.3115/v1/D14-1162}
 }
 
 @article{mikolov2013efficient,
@@ -43,8 +44,9 @@
 }
 
 @article{krijthe2015rtsne,
-  title={Rtsne: T-distributed stochastic neighbor embedding using Barnes-Hut implementation},
-  author={Krijthe, Jesse H},
-  journal={R package version 0.13, URL https://github. com/jkrijthe/Rtsne},
-  year={2015}
+  title = {Rtsne: T-distributed stochastic neighbor embedding using Barnes-Hut implementation},
+  author = {Krijthe, Jesse H},
+  note = {R package version 0.13},
+  url = {https://github.com/jkrijthe/Rtsne},
+  year = {2015}
 }
