### 2020-06-13 #2273 @labarba
👋 @karthik — Could you have a look at this submission and let us know if you can handle it as editor? Perhaps you can accept this on a 'best effort' basis (we have more than 30 submissions in the backlog).

Title: _memr: An R package for medical free-text records embedding and visualization_
