### 2020-07-15 #2482 @whedon
[REVIEW]: memr: An R package for medical free-text records embedding and visualization
**Submitting author:** @adamgdobrakowski (<a href="http://orcid.org/0000-0001-6460-3079">Adam Dobrakowski</a>)
**Repository:** <a href="https://github.com/MI2DataLab/memr" target ="_blank">https://github.com/MI2DataLab/memr</a>
**Version:** v0.1.0

@cmaimone & @ @timchurches, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
