### 2020-07-19 #2482 @karthik
Thanks for letting me know @cmaimone
I fixed a few problems with the way they reference citations and sent in a pull-request. Once merged, Whedon will be able to generate PDFs again. In the meantime here's the paper I was able to compile locally with my fix.

[nlr9oojjkffpxi0pif2v.pdf](https://github.com/openjournals/joss-reviews/files/4943009/nlr9oojjkffpxi0pif2v.pdf)
