### 2020-07-19 MI2DataLab/memr#2 @cmaimone
Package installation instructions
Consider recommending remotes::install_git() instead of devtools::install_git() as the remotes package has fewer dependencies than devtools and is less likely to cause installation issues for people.

Also consider submitting the package to CRAN or bioconductor to make discovery and installation easier for more R users.

Re: https://github.com/openjournals/joss-reviews/issues/2482
