### 2020-07-19 MI2DataLab/memr#4 @cmaimone
Contributing, issue submission, and help guidelines
re: https://github.com/openjournals/joss-reviews/issues/2482

memr needs:
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
