### 2020-07-19 MI2DataLab/memr#5 @cmaimone
JOSS paper issue
re: https://github.com/openjournals/joss-reviews/issues/2482

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

What other packages have similar functionality to this one?  Without memr, what would a researcher or doctor use to perform similar analysis on their data?  What does memr add over someone using text2vec and performing cluster or PCA analysis themselves?
