### 2020-07-19 MI2DataLab/memr#6 @cmaimone
JOSS paper citations
Re: re: https://github.com/openjournals/joss-reviews/issues/2482

The JOSS paper should cite the other R packages you use.  It currently only cites text2vec.
