### 2020-07-19 #2482 @cmaimone
@karthik I'm not sure this submission meets the substantial scholarly effort criteria

> Age of software (is this a well-established software project) / length of commit history.
github history is just a few months

> Number of commits.
15

> Number of authors.
1

> Lines of code (LOC): These statistics are usually reported by Whedon in the pre-review issue thread.
919 lines of R code, under the recommended 1000

> Whether the software has already been cited in academic papers.
No evidence of this beyond the author's preprint https://arxiv.org/pdf/1907.04152.pdf

> Whether the software is sufficiently useful that it is likely to be cited by other researchers working in this domain.
The package does not appear to have sufficient documentation and examples to lead to widespread use that would bring about citations.  The package accompanies https://arxiv.org/pdf/1907.04152.pdf, but it does not go beyond serving as a record of the code and methods used for that paper.
