### 2020-07-23 #2482 @adamgdobrakowski
@cmaimone Thank you for reviewing my package. I will try to solve all the opened issues. I would like also to answer to some of your doubts:

>  github history is just a few months
>  Number of commits.
>    15

The tools included in the package had been developing since 2018, but they were stored in a private repository due to medical data sensitivity and privacy. First experiments were performed on real medical data, and the package could be transformed into a public repository after a separation from the real data. Of course there are few commits because an initial commit was with a complete package.

>  Whether the software has already been cited in academic papers.
> No evidence of this beyond the author's preprint https://arxiv.org/pdf/1907.04152.pdf

My paper is accepted at the conference ISMIS 2020 https://ismis.ist.tugraz.at/accepted-papers/ and I will change the citation when it will be published.

> The package does not appear to have sufficient documentation and examples to lead to widespread use that would bring about citations.

I will develop a documentation and I could give more examples with a new bigger artificial dataset.
