### 2020-07-24 #2482 @timchurches
I'm a bit confused as to what I am supposed to do at this point... I've looked at the 4 page manuscript, and I agree with @cmaimone that the project is insufficiently mature and insufficiently documented to warrant review as a software package at this stage. The manuscript as it stands is really more a research note, explaining how a known body of methods (word embedding etc) can be applied to a particular domain (medical free text), although it must e noted that there are dozens of such papers already. I suggest that the manuscript be rejected with an invitation to resubmit when the project is more mature and complete, or to submit elsewhere as a research note type of paper. I am presuming that the peer review process for JOSS is just that, a review process, rather than a mentoring process intended to help the manuscript author(s) develop their project to the point that it meets the criteria for JOSS consideration. If mentoring is intended, it might be done better through organisations such as rOpenSci, rather than via an extended manuscript peer review process.
