### 2020-07-24 #2482 @timchurches
I should add that even as a research methodology or research note paper, the manuscript still needs a lot of work. Such a paper needs to assume the reader is unfamiliar with the methods used, and thus needs to explain, possibly in non-mathematical terms, what embedding is, how similarity between or clustering of embedding vectors can be used for tasks relevant to researchers working with free text records, what metrics might be used to evaluate this, and explanations that, for example, that tSNE is a way of visualising clusters in a reduce number of dimensions, and so on. The manuscript does none of those things at present. Nor does it give details of the software package or its design and architecture. Thus, unless JOSS reviews are intended to be a mentoring service, suggest rejection with fresh submission of a much expanded manuscript at a later date if the author wishes.
