### 2020-07-24 #2482 @timchurches
Sorry, last post on this: I just read the paper referenced in the manuscript: https://arxiv.org/abs/1907.04152 That looks much more like the research note or research methodology type of paper I described above. If it has not been submitted elsewhere, that paper would be a much better basis for a submission to JOSS, if a section or two on the software package were added to it. But that needs to be a separate submission with the additional authors named in the pre-print, not a revision of the manuscript under consideration here.
