### 2020-07-24 #2482 @timchurches
A final, more general point: I notice that the JOSS review questions does not include anything about ethics oversight. Obviously where a manuscript reports results from the software package in question which are obtained using data or information that needs to be under human research ethics committee (IRB) oversight, the details of the necessary clearances need to be provided in the manuscript. The preprint cited in the submitted manuscript describes the use of free text notes extracted from 100,000 patient records. In most jurisdictions, such use would require ethics review and approval, but none was mentioned. Just something to bear in mind when reviewing JOSS manuscripts that deal with software used for human and/or animal research, where examples of such research are being reported in the JOSS paper. I'm sure the authors of the preprint took all the required steps for ethical approval of their work in their jurisdiction, they just need to mention that in any paper submitted to JOSS for consideration.
