### 2020-07-27 #2482 @karthik
@cmaimone and @timchurches , thank you for the review and your thoughts on the paper. I note both of you concurring that this is not substantive enough to count as a scholarly effort for JOSS and I also see the additional concerns raised by @timchurches

>  I suggest that the manuscript be rejected with an invitation to resubmit when the project is more mature and complete, or to submit elsewhere as a research note type of paper.

> I am presuming that the peer review process for JOSS is just that, a review process, rather than a mentoring process intended to help the manuscript author(s) develop their project to the point that it meets the criteria for JOSS consideration. If mentoring is intended, it might be done better through organisations such as rOpenSci, rather than via an extended manuscript peer review process.

You presume correct. JOSS does not mentor papers into completion. Editors at rOpenSci will have to decide if this is a fit, but for JOSS we just make a decision based on what was submitted (with minor revisions).

>  I notice that the JOSS review questions do not include anything about ethics oversight. Obviously where a manuscript reports results from the software package in question which are obtained using data or information that needs to be under human research ethics committee(HREC aka IRB) oversight, the details of the necessary clearances need to be provided in the manuscript.

Thank you. I'll raise this issue with the JOSS editorial board.

@arfon @kthyng Please see discussion above. My recommendation supports the view of the reviewers, that this be rejected from JOSS in the current form. The authors can submit a presubmission inquiry in the future to determine if the paper/software meets the criteria for acceptance to JOSS.
