### 2020-08-04 #2482 @karthik
:wave: @adamgdobrakowski

Thank you for submitting your software to JOSS. After considering the comments of the two reviewers and discussion with the EICs, I have decided to reject this submission in the current form.

We are rejecting this submission because:

-  It currently does not meet our criteria for a substantial scholarly effort
- There isn't sufficient indication that your software is will be of interest to the broader research community
- There are unanswered questions about ethics oversight and missing IRB details.

If you are able to address these in a revision and demonstrate that it is a substantial scholarly effort, please submit a pre-submission inquiry and reference this review thread. We wish you the best.
